Expose plugin API via window
