+++ frontend/packages/console-dynamic-plugin-sdk/src/runtime/.eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  root: true,
+  extends: ['plugin:console/node-typescript-prettier'],
+};
