+++ frontend/packages/console-dynamic-plugin-sdk/src/runtime/plugin-api.ts
@@ -1,0 +1,8 @@
+export const exposePluginAPI = () => {
+  window.api = {
+    useK8sWatchResource: require('@console/internal/components/utils/k8s-watch-hook')
+      .useK8sWatchResource,
+    useK8sWatchResources: require('@console/internal/components/utils/k8s-watch-hook')
+      .useK8sWatchResources,
+  };
+};
